VideoCore: Move LookupTexture out of debug_utils.h
