+++ src/video_core/CMakeLists.txt
@@ -1,36 +1,29 @@
 set(SRCS
+            clipper.cpp
+            command_processor.cpp
+            debug_utils/debug_utils.cpp
+            pica.cpp
+            primitive_assembly.cpp
+            rasterizer.cpp
+            renderer_base.cpp
             renderer_opengl/gl_rasterizer.cpp
             renderer_opengl/gl_rasterizer_cache.cpp
             renderer_opengl/gl_shader_gen.cpp
             renderer_opengl/gl_shader_util.cpp
             renderer_opengl/gl_state.cpp
             renderer_opengl/renderer_opengl.cpp
-            debug_utils/debug_utils.cpp
-            clipper.cpp
-            command_processor.cpp
-            pica.cpp
-            primitive_assembly.cpp
-            rasterizer.cpp
-            renderer_base.cpp
             shader/shader.cpp
             shader/shader_interpreter.cpp
             swrasterizer.cpp
+            texture/texture_decode.cpp
             vertex_loader.cpp
             video_core.cpp
             )
 
 set(HEADERS
-            debug_utils/debug_utils.h
-            renderer_opengl/gl_rasterizer.h
-            renderer_opengl/gl_rasterizer_cache.h
-            renderer_opengl/gl_resource_manager.h
-            renderer_opengl/gl_shader_gen.h
-            renderer_opengl/gl_shader_util.h
-            renderer_opengl/gl_state.h
-            renderer_opengl/pica_to_gl.h
-            renderer_opengl/renderer_opengl.h
             clipper.h
             command_processor.h
+            debug_utils/debug_utils.h
             gpu_debugger.h
             pica.h
             pica_state.h
@@ -39,10 +32,19 @@ set(HEADERS
             rasterizer.h
             rasterizer_interface.h
             renderer_base.h
+            renderer_opengl/gl_rasterizer.h
+            renderer_opengl/gl_rasterizer_cache.h
+            renderer_opengl/gl_resource_manager.h
+            renderer_opengl/gl_shader_gen.h
+            renderer_opengl/gl_shader_util.h
+            renderer_opengl/gl_state.h
+            renderer_opengl/pica_to_gl.h
+            renderer_opengl/renderer_opengl.h
             shader/debug_data.h
             shader/shader.h
             shader/shader_interpreter.h
             swrasterizer.h
+            texture/texture_decode.h
             utils.h
             vertex_loader.h
             video_core.h
